From: Martin Wilck <mwilck@suse.com> Subject: [PATCH 23/26] autoconf: remove support for internal talloc and  libupnp
The internal versions of libupnp and talloc included in djmount
are outdated and shouldn't be used any more.

The "2GB" problem looks like a generic problem with libupnp still
in current versions, but using our own internal library version is not
the way to go. We need to find a different solution.